fix : 비밀번호, 로그인, 회원가입 코드 수정

+++ src/main/java/com/douzone/wehago/domain/User.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @AllArgsConstructor
 public class User {
 
-    private Long userSeq;
+    private Integer userSeq;
     private String userId;
     private String userPwd;
     private String userName;
